implementacion de control de dialogo

+++ cedesistemas-app/src/app/users/list/list.component.ts
@@ -2,6 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import { MatTableDataSource } from '@angular/material/table';
 import { UserModel } from '../user.model';
 import { UsersService } from '../users.service';
+import { MatDialog } from '@angular/material/dialog';
+import { DialogComponent } from 'src/app/shared/dialog/dialog.component';
+import { TranslateService } from '@ngx-translate/core';
 
 @Component({
   selector: 'app-list',
@@ -9,40 +12,61 @@ import { UsersService } from '../users.service';
   styleUrls: ['./list.component.scss']
 })
 export class ListComponent implements OnInit {
-  displayedColumns: string[] = ['FullName', 'Nit', 'BirthDay', 'Ranking','IsAdmin', 'Edit', 'Delete'];
+  displayedColumns: string[] = ['FullName', 'Nit', 'BirthDay', 'Ranking', 'IsAdmin', 'Edit', 'Delete'];
   dataSource: MatTableDataSource<UserModel>;
-  private isLoading:boolean= true;
+  private isLoading: boolean = true;
   applyFilter(filterValue: string) {
     this.dataSource.filter = filterValue.trim().toLowerCase();
   }
-  constructor(private usersService: UsersService) { }
+  constructor(private usersService: UsersService, private dialog: MatDialog, private translate: TranslateService) { }
 
   ngOnInit() {
     this.load();
   }
-  load(){
+  load() {
     this.usersService.get().subscribe(
-      (data:UserModel[])=>{ 
-        this.isLoading=false;
+      (data: UserModel[]) => {
+        this.isLoading = false;
         this.dataSource = new MatTableDataSource(data);
       },
-      (error:any)=>{
-          this.isLoading=false;
-          alert(error)
+      (error: any) => {
+        this.isLoading = false;
+        alert(error)
       }
     );
   }
-  delete(element:UserModel){ 
-      this.usersService.delete(element.Id).subscribe(
-        (data)=>{
-          this.isLoading=false;
-          this.load();
-        },
-        (error)=>{
-          this.isLoading=false;
-          alert(error)
-        } 
-      );
+  delete(element: UserModel) {
+
+
+    this.translate.get("User.DeleteMessage").subscribe((msg:string)=>{
+
+      let dialogResponse = this.dialog.open(DialogComponent, {
+        data: {
+          title: element.Name + " " + element.LastName, 
+          message: msg
+        }
+      });
+  
+      dialogResponse.afterClosed().subscribe(result=>{
+        if(result){
+          this.usersService.delete(element.Id).subscribe(
+            (data) => {
+              this.isLoading = false;
+              this.load();
+            },
+            (error) => {
+              this.isLoading = false;
+              alert(error)
+            }
+          );
+        }
+      });
+
+    });
+
+
+   
+
   }
 
 }
